@@ -2,8 +2,8 @@ package io.github.nkrusch.spacelaunchone.features;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.RecyclerView;
 import android.view.Menu;
 import android.view.MenuItem;
 
