@@ -2,7 +2,7 @@ package service;
 
 import android.app.IntentService;
 import android.content.Intent;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import api.OnLoadCallback;
 
